server-python/requirements.txt
Flask
flask-cors
python-dotenv
google-generativeai
google-genai